Move GrpcMethod inside GrpcClient
